refactor(srv): move reducer.rs's monolithic test module into per-domain submodule tests
Tier 1 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
93% of reducer.rs (~3750 of 4044 lines) was a single flat #[cfg(test)] mod
tests covering every command domain, even though production code was already
cleanly split into mod block/layout/lifecycle/snapshot/tab/window/workspace
submodules with zero tests of their own. Moved each domain's tests into its
own submodule file, plus a new reducer/test_support.rs for shared fixtures
(ctx, create_workspace/tab/block, etc.). reducer.rs shrinks to 316 lines
(mod declarations, Ctx, update(), one genuinely cross-domain test for the
catch-all dispatch arm).

Pure relocation, no assertion changes. Test count unchanged: 120 passed
before and after in reducer::, full crate suite 1639 passed / 0 failed.

<!-- agentmux:agent_id=agentx -->
