+++ agentmux-srv/src/reducer/block.rs
@@ -166,3 +166,205 @@ pub(super) fn handle_update_block_meta(
         version: v,
     }]
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::reducer::test_support::*;
+    use crate::reducer::update;
+    use agentmux_common::ipc::Command;
+
+    #[test]
+    fn create_block_validates_tab_exists() {
+        let mut state = State::default();
+        let events = update(
+            &mut state,
+            Command::CreateBlock { tab_id: "no-such-tab".into(), meta: serde_json::Value::Null },
+            &ctx(1),
+        );
+        assert!(matches!(&events[0], Event::Error { .. }));
+        assert!(state.blocks.is_empty());
+    }
+
+    #[test]
+    fn create_block_appends_to_tab_block_ids() {
+        let mut state = State::default();
+        let ws_id = create_workspace(&mut state, "w");
+        let tab_id = create_tab(&mut state, &ws_id, "t");
+        let events = update(
+            &mut state,
+            Command::CreateBlock { tab_id: tab_id.clone(), meta: serde_json::Value::Null },
+            &ctx(2),
+        );
+        assert!(matches!(&events[0], Event::BlockCreated { .. }));
+        let block_id = match &events[0] {
+            Event::BlockCreated { block_id, .. } => block_id.clone(),
+            _ => panic!(),
+        };
+        assert_eq!(state.tabs[&tab_id].block_ids, vec![block_id.clone()]);
+        assert_eq!(state.blocks[&block_id].tab_id, tab_id);
+    }
+
+    #[test]
+    fn delete_block_removes_from_state_and_tab() {
+        let mut state = State::default();
+        let ws_id = create_workspace(&mut state, "w");
+        let tab_id = create_tab(&mut state, &ws_id, "t");
+        let _ = update(
+            &mut state,
+            Command::CreateBlock { tab_id: tab_id.clone(), meta: serde_json::Value::Null },
+            &ctx(2),
+        );
+        let block_id = state.tabs[&tab_id].block_ids[0].clone();
+        let events = update(
+            &mut state,
+            Command::DeleteBlock {
+                tab_id: tab_id.clone(),
+                block_id: block_id.clone(),
+            },
+            &ctx(3),
+        );
+        assert!(matches!(&events[0], Event::BlockDeleted { .. }));
+        assert!(!state.blocks.contains_key(&block_id));
+        assert!(state.tabs[&tab_id].block_ids.is_empty());
+    }
+
+    #[test]
+    fn delete_block_unknown_silent_no_op() {
+        let mut state = State::default();
+        let ws_id = create_workspace(&mut state, "w");
+        let tab_id = create_tab(&mut state, &ws_id, "t");
+        let events = update(
+            &mut state,
+            Command::DeleteBlock {
+                tab_id,
+                block_id: "ghost".into(),
+            },
+            &ctx(2),
+        );
+        assert!(events.is_empty());
+    }
+
+    // ---- Phase E.5.5 — MoveBlock tests ----
+
+    #[test]
+    fn move_block_cross_tab_updates_lists_and_parent() {
+        let mut state = State::default();
+        let ws = create_workspace(&mut state, "w");
+        let src_tab = create_tab(&mut state, &ws, "src");
+        let dst_tab = create_tab(&mut state, &ws, "dst");
+        let block = create_block(&mut state, &src_tab);
+        let dst_existing = create_block(&mut state, &dst_tab);
+        let events = update(
+            &mut state,
+            Command::MoveBlock {
+                block_id: block.clone(),
+                src_tab_id: src_tab.clone(),
+                dst_tab_id: dst_tab.clone(),
+                dst_index: 0,
+            },
+            &ctx(2),
+        );
+        assert!(matches!(&events[0], Event::BlockMoved { .. }));
+        assert_eq!(state.tabs[&src_tab].block_ids, Vec::<String>::new());
+        assert_eq!(state.tabs[&dst_tab].block_ids, vec![block.clone(), dst_existing]);
+        assert_eq!(state.blocks[&block].tab_id, dst_tab);
+    }
+
+    #[test]
+    fn move_block_intra_tab_repositions() {
+        let mut state = State::default();
+        let ws = create_workspace(&mut state, "w");
+        let tab = create_tab(&mut state, &ws, "t");
+        let b1 = create_block(&mut state, &tab);
+        let b2 = create_block(&mut state, &tab);
+        let b3 = create_block(&mut state, &tab);
+        // Move b1 to position 2 (end after removal).
+        let events = update(
+            &mut state,
+            Command::MoveBlock {
+                block_id: b1.clone(),
+                src_tab_id: tab.clone(),
+                dst_tab_id: tab.clone(),
+                dst_index: 2,
+            },
+            &ctx(2),
+        );
+        match &events[0] {
+            Event::BlockMoved { dst_index, .. } => assert_eq!(*dst_index, 2),
+            other => panic!("expected BlockMoved, got {:?}", other),
+        }
+        assert_eq!(state.tabs[&tab].block_ids, vec![b2, b3, b1]);
+    }
+
+    #[test]
+    fn move_block_rejects_unknown_src_or_dst_or_block() {
+        let mut state = State::default();
+        let ws = create_workspace(&mut state, "w");
+        let tab = create_tab(&mut state, &ws, "t");
+        let other_tab = create_tab(&mut state, &ws, "other");
+        let block = create_block(&mut state, &tab);
+
+        let events = update(
+            &mut state,
+            Command::MoveBlock {
+                block_id: block.clone(),
+                src_tab_id: "ghost-src".into(),
+                dst_tab_id: other_tab.clone(),
+                dst_index: 0,
+            },
+            &ctx(2),
+        );
+        assert!(matches!(&events[0], Event::Error { .. }));
+
+        let events = update(
+            &mut state,
+            Command::MoveBlock {
+                block_id: block.clone(),
+                src_tab_id: tab.clone(),
+                dst_tab_id: "ghost-dst".into(),
+                dst_index: 0,
+            },
+            &ctx(3),
+        );
+        assert!(matches!(&events[0], Event::Error { .. }));
+
+        let events = update(
+            &mut state,
+            Command::MoveBlock {
+                block_id: "ghost-block".into(),
+                src_tab_id: tab,
+                dst_tab_id: other_tab,
+                dst_index: 0,
+            },
+            &ctx(4),
+        );
+        assert!(matches!(&events[0], Event::Error { .. }));
+    }
+
+    #[test]
+    fn move_block_rejects_when_block_belongs_to_different_tab() {
+        let mut state = State::default();
+        let ws = create_workspace(&mut state, "w");
+        let real_src = create_tab(&mut state, &ws, "real");
+        let other = create_tab(&mut state, &ws, "other");
+        let dst = create_tab(&mut state, &ws, "dst");
+        let block = create_block(&mut state, &real_src);
+        let events = update(
+            &mut state,
+            Command::MoveBlock {
+                block_id: block,
+                src_tab_id: other,
+                dst_tab_id: dst,
+                dst_index: 0,
+            },
+            &ctx(2),
+        );
+        match &events[0] {
+            Event::Error { message, .. } => {
+                assert!(message.contains("belongs to tab"), "got: {}", message);
+            }
+            other => panic!("expected Error, got {:?}", other),
+        }
+    }
+}
